validate transition
