+++ frontend/src/component/project/Project/Import/ImportStage.ts
@@ -1,0 +1,4 @@
+export type ImportStage =
+    | { name: 'configure' }
+    | { name: 'validate'; environment: string; payload: string }
+    | { name: 'import' };
